move resblock and spatial-T to common.hpp to reuse in control.hpp
